Rename to `pr-from-branch`
