+++ readme.md
@@ -100,7 +100,7 @@ https://github.com/refined-github/refined-github/wiki/Contributing#metadata-guid
 - [](# "github-actions-indicators") [In the workflows sidebar, shows an indicator that a workflow can be triggered manually, and its next scheduled time if relevant.](https://github-production-user-asset-6210df.s3.amazonaws.com/83146190/252181237-a1d809b1-e5d4-4747-9654-7dde5ab5c61a.png)
 - [](# "quick-repo-deletion") [Simplifies the deletion of repositories. Adds "Delete fork" button on 0-star repos. Enables <kbd>shift</kbd> <kbd>alt</kbd> clicks to delete repos in 2 clicks.](https://github.com/user-attachments/assets/9e05ec18-680d-4fbd-acbd-2b5b3505c5b5)
 - [](# "delete-branch") [Adds a button to delete the current branch on the branch page.](https://github.com/user-attachments/assets/02478585-0613-4070-9616-df069dbc6c38)
-- [](# "unwrap-contribute") [Lets you open a PR in one click on the branch page.](https://github.com/user-attachments/assets/7f42664a-76e6-4518-88b2-20b69a3260f9)
+- [](# "pr-from-branch") [Lets you open a PR in one click on the branch page.](https://github.com/user-attachments/assets/7f42664a-76e6-4518-88b2-20b69a3260f9)
 - [](# "archive-forks-link") [Helps you find forks on archived repos.](https://user-images.githubusercontent.com/1402241/230362566-12493c80-bffe-4c7a-b9ba-4a11b1358ab0.png)
 - [](# "repo-avatars") [Adds the profile picture to the header of public repositories.](https://github-production-user-asset-6210df.s3.amazonaws.com/1402241/255323568-aee4d90e-844e-41e8-880a-ce466826516c.png)
 - [](# "small-user-avatars") [Shows a small avatar next to the username in issue/PR lists and mentions.](https://github-production-user-asset-6210df.s3.amazonaws.com/1402241/271184107-24ec471e-54d1-434a-a5f2-615902d2cad9.png)
